+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/cache_dsn.php
@@ -1,0 +1,15 @@
+<?php
+
+$container->loadFromExtension('framework', [
+    'cache' => [
+        'app' => 'redis://localhost',
+        'pools' => [
+            'cache.custom_dsn' => [
+                'adapter' => 'memcached://localhost',
+            ],
+            'cache.env_dsn' => [
+                'adapter' => '%env(CACHE_DSN)%',
+            ],
+        ],
+    ],
+]);